backend/requirements.txt
amqp==2.6.1
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==4.4.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==3.2.12
django-auto-prefetching==0.2.11
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-import-export==2.8.0
django-rest-multiple-models==2.1.3
django-simple-history==3.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
et-xmlfile==1.1.0
gunicorn==20.1.0
kombu==4.6.11
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
Pillow==9.1.0
prompt-toolkit==3.0.30
PyJWT==2.3.0
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
redis==4.3.4
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
vine==1.3.0
wcwidth==0.2.5
whitenoise==6.0.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
